# hospital-surgeons/public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Hospital Surgeons API",
    "version": "1.0.0",
    "description": "API documentation for Hospital Surgeons application",
    "contact": {
      "name": "API Support",
      "email": "admin@example.com"
    }
  },
  "servers": [
    {
      "url": "https://hospital-surgoen.onrender.com",
      "description": "API Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT token"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/hospitals/{id}/dashboard": {
      "get": {
        "summary": "Get hospital dashboard metrics and data",
        "tags": [
          "Hospitals"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Hospital ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Dashboard data retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "metrics": {
                          "type": "object"
                        },
                        "todaysSchedule": {
                          "type": "array"
                        },
                        "recentAssignments": {
                          "type": "array"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}